# Add an index to better support `smaller_number_is_higher_priority`